apps/doc-ingest/src/fn/requirements.txt
parliament-functions==0.1.0
boto3==1.26.16
langchain==0.0.324
pymilvus==2.3.1
bs4==0.0.1
sentence-transformers==2.2.2